% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLBteams-package.R
\docType{package}
\name{MLBteams-package}
\alias{MLBteams}
\alias{MLBteams-package}
\title{MLBteams: Creates Data For MLB Team Names And Stadiums}
\description{
This package creates a re-usable data from the wikipedia page on MLB stadiums. The current version uses the information available prior to the 2021 season.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seantoconnor/MLBteams}
}

}
\author{
\strong{Maintainer}: Sean O'Connor \email{stoconno09@gmail.com}

}
\keyword{internal}
